IntelliAI/requirements.txt
agno
milvus